Mj/file store err flip (#1074)
* Add specific Error for handling timestamps

`TimestampDecodeError` provides more explicit errors for what type of 
field we are attempting to decode, and an extra if we cannot parse the 
incoming number.

`TimestampDecodeResult` is provided as a convenience for implementing 
MsgTimestamp.

```
impl MsgTimeStamp<Result<DateTime<Utc>, TimestampDecodeError>> for Type {}
```
is super cumbersome to read.

* Raise constraints on MsgDecode + Custom Error

This will cuase the compiler to err if you try to implement MsgDecoe for
a type without also having the TryFrom impl.

Before, that error would only be thrown if someone tried to call 
::decode() on a msg.

We also provide a custom decode error, this allows us to tell the 
different between prost failing to decode a protobuf vs. our own 
conversion code failing to transform into a domain struct.

This also allows us to provide our own error for conversions instead of 
being tied to `file_store::Error` thta had become a grab bag of domain 
specific err variants.

* FileInfoPollerParser allow user errors to pass through

Remove the constraint on MsgDecode implementers returning a 
`file_store::Error` in lieu of their own domain error. The parser will 
pass them back their own error if something fails, and they can be more 
explicit about the errors they care about. 

Rather than trying to decide between putting a new err variant in 
file-store, or using ExternalError, or putting a specific string in the 
wrong err type for what they actual are trying to communicate.

* Converting from a IotBeaconReport to a regular beacon cannot fail

* mapper to go from prost i32 into enum or err

Use the member field type to figure out what enum we’re casting into. 
Provide an err variant that wraps a `prost::UnknownEnumValue` for 
propogating unsupported variants in TryFrom impls.

* Cleanup file-store errors to only contain errors that can be returned from file-store lib code

All domain specific errors are going to move into the err variant of
TryFrom for the type that is being converted to.

* Create domain errors for iot proto types

Includes the normal `TimestampDecodeError` provided by file-store. 
MissingField when trying to access nested types to convert. 

And wrapping `prost::UnknownEnumValue` errors to better know which enum 
we could not map, and have a consistent way of treating enum 
conversions.

Before we had a smattering of using `Enum::try_from(proto.enum_field)` 
which allowed is to catch errs. And `proto.enum_field()` using the 
provided getter that will panic if the value is unknown. 

Now we have a `prost_enum(proto.enum_field, Error::EnumField)` mapper 
being provided that can figure out what type we’re trying to convert 
into and forward any errors back to us for bubbling.

* Create domain errors for mobile proto types

* Create domain errors for subscriber proto types

You may notice an `h3o::error::InvalidCellIndex` error in usage counts.

One of the benefits for flipping error logic over to the implementer is 
now file-store has no more reason to know about h3o.

* Add domain errors for conversions outside file-store-oracles

These 2 files are essentially the same and are good candidates for 
moving into networ_common in file-store-oracles. That is outside the 
scope of this PR.

* Use a more aligned error for testing

* Handling reward manifest still require error mapping

Many of the streams in file-store are still heavily tied to 
`file_store::Error`. Passing reward manifest files still requires us 
mapping into a proper lib error so we can pass it to a lib function.

* Sort deps for niceness, and cargo.lock update