Unignore as many tests as we can (#1093)
These tests were ignored when CI stopped being able to complete complaining about running out of memory. The first solution was to clear the build cache of containers we were using to run tests in. Another solution that we never got to was increasing the shared memory size for the postgres container.